[fix] rand関数のキャストを修正

+++ Bve5_Parsing/MapGrammar/AstEvaluator.cs
@@ -316,7 +316,7 @@ namespace Bve5_Parsing.MapGrammar
                 return random.NextDouble();
 
 
-            return random.Next((int)Visit(node.Value));
+            return random.Next((int)((double)Visit(node.Value)));
         }
 
         /// <summary>
